1、修复些计算性上的错误。 2、floor在有些平台无法编译的问题 3、增加某些阴历转阳历，阳历转阴历的方法函数

+++ src/JD.cpp
@@ -24,7 +24,7 @@ long double JD::DD2JD(int y, uint8_t m, long double d)
 }
 
 //儒略日数转公历
-Time JD::JD2DD(int jd)
+Time JD::JD2DD(long double jd)
 {
 	Time r;
 	int D = int2(jd + 0.5);
